refactor(phase-3a): consolidate customer data access, drop sync getDb
- Move createOrRetrieveCustomerId from lib/customer.ts into lib/db/customer.ts
  with an object signature and optional prisma transaction param; delete
  lib/customer.ts and update the single checkout caller
- Remove the synchronous getDb() from prisma.ts; migrate its 4 page call
  sites to await getDbAsync() (memoization left untouched)

Logic bodies byte-identical; no external API (v3) touched. lint + tsc green.
