+++ src/prisma.ts
@@ -2,22 +2,6 @@ import { PrismaClient } from "@prisma/client";
 import { PrismaPg } from "@prisma/adapter-pg";
 import { getCloudflareContext } from "@opennextjs/cloudflare";
 
-export const getDb = () => {
-  const { env } = getCloudflareContext({ async: false });
-
-  if (!env.BEUTL_DATABASE_HYPERDRIVE) {
-    throw new Error("BEUTL_DATABASE_HYPERDRIVE binding not found");
-  }
-
-  const connectionString = env.BEUTL_DATABASE_HYPERDRIVE.connectionString;
-  if (!connectionString) {
-    throw new Error("Hyperdrive connection string not available");
-  }
-
-  const adapter = new PrismaPg({ connectionString, maxUses: 1 });
-  return new PrismaClient({ adapter });
-};
-
 // For Server Actions and other server-side contexts
 export const getDbAsync = async () => {
   const { env } = await getCloudflareContext({ async: true });
